requirements.txt
Brotli==1.0.9
click==8.1.3
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.13
ipykernel==6.15.3
ipywidgets==8.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
plotly==5.10.0
pydeck==0.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
tenacity==8.1.0
traitlets==5.4.0
Werkzeug==2.2.2

# Local package in ./src
./src/dash_deck-0.0.2.tar.gz

# Required for deployment on Heroku
gunicorn==20.1.0